refactor(segmentation): rename instance segmenter
