+++ examples/segmentation/CMakeLists.txt
@@ -1,2 +1,2 @@
-add_subdirectory(instance-segmenter/src/cpp instance-segmenter)
+add_subdirectory(single-stream-instance-segmenter/src/cpp single-stream-instance-segmenter)
 add_subdirectory(yolov8-instance-segmenter/src/cpp yolov8-instance-segmenter)
