+++ examples/segmentation/single-stream-instance-segmenter/src/common/coco_label.txt
