+++ examples/segmentation/single-stream-instance-segmenter/src/cpp/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.16)
 
 if (CMAKE_SOURCE_DIR STREQUAL CMAKE_CURRENT_SOURCE_DIR)
-  project(instance-segmenter LANGUAGES CXX)
+  project(single-stream-instance-segmenter LANGUAGES CXX)
   set(CMAKE_CXX_STANDARD 20)
   set(CMAKE_CXX_STANDARD_REQUIRED ON)
   set(CMAKE_CXX_EXTENSIONS OFF)
@@ -9,4 +9,4 @@ if (CMAKE_SOURCE_DIR STREQUAL CMAKE_CURRENT_SOURCE_DIR)
 endif()
 
 include("${CMAKE_CURRENT_LIST_DIR}/../../../../../cmake/ExampleModule.cmake")
-sima_neat_apps_module("instance-segmenter" UNIT_TEST E2E_TEST)
+sima_neat_apps_module("single-stream-instance-segmenter" UNIT_TEST E2E_TEST)
