+++ examples/segmentation/single-stream-instance-segmenter/src/python/main.py
