+++ examples/segmentation/single-stream-instance-segmenter/tests/test-scope.yaml
